Refactor reducers’s file names under root data node
Replace singulars with plurals.
