+++ src/data/interfaces/modal/reducers.js
