+++ src/screens/Surveys/screens/SurveyDetails/index.js
@@ -3,7 +3,7 @@ import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
 
 import * as modalActions from '../../../../data/interfaces/modal/actions';
-import { getModal } from '../../../../data/interfaces/modal/reducer';
+import { getModal } from '../../../../data/interfaces/modal/reducers';
 import * as commonActions from '../../actions';
 
 // Peer dependencies
